finish document class

+++ src/main/java/HashLib/CalcHashEngine/HashDefault.java
@@ -7,7 +7,7 @@ import java.io.IOException;
 
 public class HashDefault extends HashEngine {
     @Override
-    public int generateHashCode(Object key) throws IOException {
+    public long generateHashCode(Object key) throws IOException {
         byte[] bytes = ByteArray.toBytesStream(key);
         return HashFunctions.Polynomial(bytes);
     }
