+++ src/main/java/HashLib/CalcHashEngine/HashEngine.java
@@ -3,5 +3,5 @@ package HashLib.CalcHashEngine;
 import java.io.IOException;
 
 public abstract class HashEngine {
-    public abstract int generateHashCode(Object key) throws IOException;
+    public abstract long generateHashCode(Object key) throws IOException;
 }
